Refactor + update injection
