+++ tests/fixtures/link/absolute-dep.js
