+++ tests/fixtures/link/no-head.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+
+<body>
+    <h1>Hello, world!</h1>
+</body>
+<script type="module">
+    import { hello } from './absolute-dep.js';
+</script>
+
+</html>
