+++ tests/snapshots/link/no-head.snapshot.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html>
+<link rel="modulepreload" href="/absolute-dep.js" />
+
+
+<body>
+    <h1>Hello, world!</h1>
+</body>
+<script type="module">
+    import { hello } from './absolute-dep.js';
+</script>
+
+</html>
